eliminate some non-essential words from being loaded into flash.  multitasker and disassembler are among those to go.   Also remove one layer of flash cornerstone to avoid wasting flash due to padding out to flash block boundary.
All this because of filling up 64K of flash.  The mecrisp image presently being used was built to
run on a smaller part, and need to be recompiled and reloaded on the device to use all 128KB of
flash space actually available.  That will eventually be done, but not today..