Performance tweaks to SPI data transfer to TFT ILI9341 display.  Mostly:
- don't have the >spi word call >spi> and drop the result.  This allows
the serialization of a byte to proceed in parallel with computing the next
byte and writing it to register once the flag is ready. Previously, the
>spi> word assume the SPI device is idle, transmits a byte and waits
for the input byte to completely arrive before proceeding.

- tweak -spi word to wait for the output to drain before releasing
SPI CS.  This is probably not necessary.  Without doing this, the
CS line is released ~400ns after the last clock cycle.  Actually
polling and waiting for the BUSY flag to release take somewhat
longer at about 1800ns, probably because of just one trip through
reading the status register and testing the bit.  Perhaps do
bitband to speed this put, but most of the time spent on the
SPI interface is doing bulk multi byte transfers.

- fix setwindow word to not release/reassert CS between the
4 data bytes following the setrow and setcolumn commands used
to setup a window that a bitmap will be rendered into.  This
now transmits 4 bytes of parameters with no idle between the
bytes.

The effect of all the changes is to more than double the speed
at which the display can be cleared.